allow nvme only setup

+++ modules/install_steps.sh
@@ -169,8 +169,10 @@ install_kernel() {
     spawn_chroot "DEBIAN_FRONTEND=noninteractive apt-get -y install ${kernel_image}" || die "could not install kernel"
 
     # grub it to all disks
-    for x in /dev/sd[a-z]; do
-      spawn_chroot "/usr/sbin/grub-install $x"
+    for x in /dev/sd[a-z] /dev/nvme[0-9]; do
+      if [ -e $x ]; then
+        spawn_chroot "/usr/sbin/grub-install $x"
+      fi
     done
   fi
 }
